move ip whitelist check into execution py

+++ util/execution.py
@@ -2,6 +2,7 @@ import subprocess
 import time
 import tempfile
 import util
+import message
 
 
 #
@@ -10,9 +11,12 @@ import util
 # be allowed to access the server
 #
 def exec_command(command):
-    result_success = subprocess.check_output(
-        [command], stderr=subprocess.STDOUT, shell=True, universal_newlines=True)
-    return util.replace_all(result_success, util.get_replace_dic())
+    if util.valid_ip():
+        result_success = subprocess.check_output(
+            [command], stderr=subprocess.STDOUT, shell=True, universal_newlines=True)
+        return util.replace_all(result_success, util.get_replace_dic())
+    else:
+        return message.error_404_msg
 
 
 #
@@ -20,23 +24,26 @@ def exec_command(command):
 # the command is run in exec_time_in_seconds
 #
 def exec_command_async(command, exec_time_in_seconds):
-    try:
-        temp_out_file_name = tempfile.NamedTemporaryFile().name
-        fout = open(temp_out_file_name, "w")
-        proc = subprocess.Popen(command, shell=True, stdout=fout)
-        # wait for a few seconds
-        time.sleep(exec_time_in_seconds)
-        # stop the process
-        proc.kill()
-        fout.close()
-        fin = open(temp_out_file_name, "rb")
-        output = fin.read()
-        fin.close()
+    if util.valid_ip():
+        try:
+            temp_out_file_name = tempfile.NamedTemporaryFile().name
+            with open(temp_out_file_name, "w") as fout:
+                proc = subprocess.Popen(command, shell=True, stdout=fout)
+                # wait for a few seconds
+                time.sleep(exec_time_in_seconds)
+                # stop the process
+                proc.kill()
 
-        # remove the temp file
-        util.delete_file(temp_out_file_name)
-        return util.replace_all(output, util.get_replace_dic())
-    except subprocess.SubprocessError:
-        proc.kill()
-        outs, errs = proc.communicate()
-        return errs
+            # read contents of the file
+            with open(temp_out_file_name, "rb") as fin:
+                output = fin.read()
+
+            # remove the temp file
+            util.delete_file(temp_out_file_name)
+            return util.replace_all(output, util.get_replace_dic())
+        except subprocess.SubprocessError:
+            proc.kill()
+            outs, errs = proc.communicate()
+            return errs
+    else:
+        return message.error_404_msg
